operate/agentic_ops/requirements.txt
pydantic-ai-slim[anthropic,openai,google]>=1.0,<2
httpx>=0.27,<1
